requirements.txt
tenacity>=8.2.0
python-dotenv
requests
pycryptodome
pytz
